+++ packages/admin-ui/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New Features
+
+-   `ViewTabCount`: Add a component that renders the number of items a view holds, to trail the label of the tab that selects it ([#82247](https://github.com/WordPress/gutenberg/pull/82247)).
+
 ## 2.9.0 (2026-08-26)
 
 ### Bug Fixes